few changes

+++ src/CreateNewFile.java
@@ -9,22 +9,25 @@ public class CreateNewFile {
         try {
             System.out.println("Please insert path");
             File path = new File(reader.readLine());
-            System.out.println(path);
-
-            System.out.println("Please insert file");
+            if (!path.exists()) {
+                throw new WrongInputException("Please input a real path");
+            }
+            System.out.println("Please choose file name to create");
             File file = new File(reader.readLine());
-            System.out.println(file);
 
             String filePath = path.getName().concat("/").concat(file.getName());
             writer = new BufferedWriter(new FileWriter(filePath));
+            System.out.println("Your file was created!");
+
         }
-        catch (IOException e) {
+        catch (IOException | WrongInputException e) {
             System.out.println(e.getMessage());
         }
         try {
             writer.close();
             reader.close();
-        } catch (IOException e) {
+        }
+        catch (IOException e) {
             throw new RuntimeException(e);
         }
     }
